CONTRATCS: force success for some pointer predicates
Force success for pointer predicates that must
necessarily hold if they ever get invoked in
an assumption context. This ensures that the value sets
of pointers as as small as possible after assuming
a requires or an ensures clause and solves a performance
issue with z3. All predicates in the `cprover_contracts.c`
library now have an extra input `may_fail` controlling the
failure behaviour, DFCC instrumentation sets the `may_fail`
parameter to true or false using recursive algorithm.